+++ lib/src/widgets/widgets.dart
@@ -1,6 +1,6 @@
+export 'dialogs/dialogs.dart';
 export 'nes_button.dart';
 export 'nes_check_box.dart';
-export 'nes_confirm_dialog.dart';
 export 'nes_container.dart';
 export 'nes_controller_focus.dart';
 export 'nes_dropshadow.dart';